pyproject.toml
[tool.poetry]
name = "localllmapi"
version = "0.1.0"
description = ""
authors = ["ethanlchristensen <ethanlchristensen@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
litestar = "^2.6.2"
uvicorn = "^0.27.1"
requests = "^2.31.0"
openai = "^1.13.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
